# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Flow API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/project": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get all projects with filtering",
                "operationId": "7390c23cd3d203ce9281eb0dfc129e63",
                "parameters": [
                    {
                        "name": "filters[name]",
                        "in": "query",
                        "description": "Filter by project name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Project A"
                        }
                    },
                    {
                        "name": "filters[department]",
                        "in": "query",
                        "description": "Filter by department (EAV attribute)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "IT"
                        }
                    },
                    {
                        "name": "filters[priority]",
                        "in": "query",
                        "description": "Filter by priority (EAV attribute)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Medium"
                        }
                    },
                    {
                        "name": "filters[status]",
                        "in": "query",
                        "description": "Filter by status (e.g., ongoing, completed)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ongoing"
                        }
                    },
                    {
                        "name": "filters[created_at][gt]",
                        "in": "query",
                        "description": "Filter projects created after a specific date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "filters[name][like]",
                        "in": "query",
                        "description": "Filter by name using LIKE operator",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "%Project%"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered list of projects"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Create a new project",
                "operationId": "8f164084581b7b455cad7055bea79e7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Project Finance Report"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ongoing"
                                    },
                                    "attributes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "attribute_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Finance"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/project/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get project by ID",
                "operationId": "2fafd6f1be81669034bbcc199d66d65f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Project"
                ],
                "summary": "Update project details",
                "operationId": "64bb23af97e74747bb323d0cdf172367",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Project Finance Report"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ongoing"
                                    },
                                    "attributes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "attribute_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Department"
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Finance"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "Delete a project",
                "operationId": "76e10b616472bb2d7a09c7f77160805a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: 'Bearer {token}'",
                "bearerFormat": "OAuth2",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}